+++ CumulusMX/Stations/EcowittExtraLogFile .cs
@@ -176,13 +176,12 @@ namespace CumulusMX.Stations
 						for (int i = 1; i <= 16; i++)
 						{
 							if (FieldIndex.TryGetValue("soilmoisture ch" + i, out idx) && int.TryParse(fields[idx], out varInt)) rec.SoilMoist[i] = varInt;
-
 							if (FieldIndex.TryGetValue("soiltemp ch" + i, out idx) && decimal.TryParse(fields[idx], out varDec)) rec.SoilTemp[i] = varDec;
 						}
 					}
 					catch (Exception ex)
 					{
-						cumulus.LogErrorMessage("EcowittExtraLogFile.DataParser: Error processing Soil Moisture - " + ex.Message);
+						cumulus.LogErrorMessage("EcowittExtraLogFile.DataParser: Error processing Soil Moisture/Temperature - " + ex.Message);
 					}
 
 					// Water - unused